streamlit/requirements.txt
streamlit==1.31.1
pandas==2.2.0
snowflake-connector-python==3.7.0
